+++ backend/src/test/java/de/govhackathon/wvsvcoronatracker/WvsvCoronaTrackerApplicationTests.java
@@ -1,4 +1,4 @@
-package de.gov_hackathon.wvsv_corona_tracker;
+package de.govhackathon.wvsvcoronatracker;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
